# docs/seo-brain/openapi.v1.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "SEO Brain API",
  "version": "0.2.0"
 },
 "paths": {
...
  "/api/v1/ai/provider-kinds": {
   "get": {
    "tags": [
     "ai"
    ],
    "summary": "Provider Kinds",
    "operationId": "provider_kinds_api_v1_ai_provider_kinds_get",
    "parameters": [
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "additionalProperties": true
         },
         "title": "Response Provider Kinds Api V1 Ai Provider Kinds Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/ai/provider-configs": {
   "get": {
    "tags": [
     "ai"
    ],
    "summary": "List Provider Configs",
    "description": "Configured providers — keys are never returned (only has_key + key_hint).",
    "operationId": "list_provider_configs_api_v1_ai_provider_configs_get",
    "parameters": [
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "additionalProperties": true
         },
         "title": "Response List Provider Configs Api V1 Ai Provider Configs Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "ai"
    ],
    "summary": "Create Provider Config",
    "operationId": "create_provider_config_api_v1_ai_provider_configs_post",
    "parameters": [
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ProviderCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Create Provider Config Api V1 Ai Provider Configs Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/ai/provider-configs/{pid}": {
   "patch": {
    "tags": [
     "ai"
    ],
    "summary": "Update Provider Config",
    "operationId": "update_provider_config_api_v1_ai_provider_configs__pid__patch",
    "parameters": [
     {
      "name": "pid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Pid"
      }
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ProviderUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Update Provider Config Api V1 Ai Provider Configs  Pid  Patch"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "ai"
    ],
    "summary": "Delete Provider Config",
    "operationId": "delete_provider_config_api_v1_ai_provider_configs__pid__delete",
    "parameters": [
     {
      "name": "pid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Pid"
      }
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Delete Provider Config Api V1 Ai Provider Configs  Pid  Delete"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/ai/provider-configs/{pid}/test": {
   "post": {
    "tags": [
     "ai"
    ],
    "summary": "Test Provider Config",
    "operationId": "test_provider_config_api_v1_ai_provider_configs__pid__test_post",
    "parameters": [
     {
      "name": "pid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Pid"
      }
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Test Provider Config Api V1 Ai Provider Configs  Pid  Test Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/ai/task-routes": {
   "get": {
    "tags": [
     "ai"
    ],
    "summary": "Task Routes",
    "operationId": "task_routes_api_v1_ai_task_routes_get",
    "parameters": [
     {
      "name": "site_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Site Id"
      }
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Task Routes Api V1 Ai Task Routes Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/ai/task-routes/{task_kind}": {
   "put": {
    "tags": [
     "ai"
    ],
    "summary": "Set Task Route",
    "operationId": "set_task_route_api_v1_ai_task_routes__task_kind__put",
    "parameters": [
     {
      "name": "task_kind",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Task Kind"
      }
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RouteSet"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Set Task Route Api V1 Ai Task Routes  Task Kind  Put"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ProviderCreate": {
    "properties": {
     "name": {
      "type": "string",
      "maxLength": 60,
      "minLength": 2,
      "title": "Name"
     },
     "kind": {
      "type": "string",
      "title": "Kind"
     },
     "api_key": {
      "anyOf": [
       {
        "type": "string",
        "minLength": 4
       },
       {
        "type": "null"
       }
      ],
      "title": "Api Key"
     },
     "base_url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Base Url"
     },
     "default_model": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Default Model"
     },
     "models": {
      "anyOf": [
       {
        "items": {
         "type": "string"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Models"
     },
     "enabled": {
      "type": "boolean",
      "title": "Enabled",
      "default": true
     }
    },
    "type": "object",
    "required": [
     "name",
     "kind"
    ],
    "title": "ProviderCreate"
   },
   "ProviderUpdate": {
    "properties": {
     "name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "api_key": {
      "anyOf": [
       {
        "type": "string",
        "minLength": 4
       },
       {
        "type": "null"
       }
      ],
      "title": "Api Key"
     },
     "base_url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Base Url"
     },
     "default_model": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Default Model"
     },
     "models": {
      "anyOf": [
       {
        "items": {
         "type": "string"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Models"
     },
     "enabled": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Enabled"
     },
     "clear_key": {
      "type": "boolean",
      "title": "Clear Key",
      "default": false
     }
    },
    "type": "object",
    "title": "ProviderUpdate"
   },
...
   "RouteSet": {
    "properties": {
     "provider_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Provider Id"
     },
     "model": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Model"
     },
     "fallback_provider_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Fallback Provider Id"
     },
     "fallback_model": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Fallback Model"
     },
     "site_id": {
      "type": "string",
      "title": "Site Id",
      "default": "*"
     },
     "policy": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Policy"
     },
     "fallbacks": {
      "anyOf": [
       {
        "items": {
         "additionalProperties": true,
         "type": "object"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Fallbacks"
     }
    },
    "type": "object",
    "title": "RouteSet"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  }
 }
}